move mock_socket to conftest.
