+++ tests/exploits/cve_2025_23016_test.py
@@ -1,21 +1,10 @@
 """Unit tests for CVE-2025-23016"""
 
 import socket
-from typing import Generator
 from unittest import mock
-
-import pytest
 
 from agent import definitions
 from agent.exploits import cve_2025_23016
-
-
-@pytest.fixture
-def mock_socket() -> Generator[mock.MagicMock, None, None]:
-    with mock.patch("socket.socket") as mock_sock_class:
-        sock_inst = mock.MagicMock()
-        mock_sock_class.return_value = sock_inst
-        yield sock_inst
 
 
 def testAccept_whenFastcgiServiceDetected_shouldReturnTrue(
